requirements.txt
requests>=2.32.0
websocket-client>=1.8.0
SQLAlchemy>=2.0.35
psycopg[binary]>=3.2.1
python-dotenv>=1.0.1
